requirements.txt
scipy==1.10.1 
torch==2.0.1 
torchvision==0.15.2 
faiss-cpu==1.7.4     
transformers==4.30.2 
scikit-learn==1.2.2